Factor out configuring job managers for the different gateways into JobManagerConfiguration.
Only configure GRAM-specific stuff if gram_gateway_enabled is True.
Have enabledServices for the various job managers return the gateway as well, instead of relying on the Gateway module to do it.
This is because RSV also uses the Gateway module but does not a gateway running on the same host,
but all of the job managers do need a gateway running on the same host.

Make CondorConfiguration.setupHTCondorCEConfig error out if htcondor-ce is not installed.

Gateway is no longer separately configurable -- all configuration that uses the [Gateway] section is done in other modules.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-configure/trunk@19174 4e558342-562e-0410-864c-e07659590f8c