Remove most usages of ce_installed(); check for specific components instead (SOFTWARE-1407)
gratia module will take CE-specific actions if a gateway is installed and ANY of the batch system-specific gratia probes or gratia-probe-gram are installed.
infoservices module will take CE-specific actions if a gateway is installed and osg-info-services is installed.
squid module will take CE-specific actions if a gateway is installed and frontier-squid is installed.
cemon module has been left unchanged.

All of the required RPMs are listed as dependencies in the osg-ce metapackage.

Note that writing the osg-[local-]job-environment.conf files still uses ce_installed().
This may or may not be changed in the future...

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-configure/trunk@19502 4e558342-562e-0410-864c-e07659590f8c