Add a Makefile for easier installation
The specfile needs to do some moves and deletions in the %install section
to make up for the things distutils can't do (such as put files in /usr/sbin).
Put those in a Makefile instead so that I can run `make install' out of SVN
and get an installation that looks the same as an RPM install.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-configure/trunk@19935 4e558342-562e-0410-864c-e07659590f8c