Add RCEntry Parameter Class for ResourceCatalog.
We can replace ResourceCatalog.add_entry() (which takes too many
parameters) with creating an RCEntry and passing it to
ResourceCatalog.add_rcentry().  RCEntry also has code for validation and
normalization (e.g. splitting up strings into lists).  Most of the logic
for representing a single entry as a list of attributes has been moved
out of ResourceCatalog and into RCEntry.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-configure/trunk@21565 4e558342-562e-0410-864c-e07659590f8c