Remove domain check for email address (SOFTWARE-2271)
Python does not have a built in way of checking for the MX record of a domain, and checking for the CNAME instead (which is what socket.gethostbyname() does) sometimes results in a bogus warning. Since we can't do the check right, it's better to remove it altogether; verifying the email format should be enough.