Remove duplication in _update_lcmaps_file
For SOFTWARE-1723. MiscConfiguration._update_lcmaps_file() has some
duplicated code to modify both the original lcmaps.db file, and an
lcmaps.db.rpmnew file (if one exists) in the same way. This duplication
was removed; in addition, osg-configure will emit a log message if it is
modifying the .rpmnew file.